[Quantization] Address review feedback round 3 on FP8 sweep
Three changes from realAsma's latest review:

- nvfp4_fp8_sweep kernel: use ``scale_safe`` rather than ``scale`` in the
  per-candidate diff so the divisor and multiplier match. Numerically
  equivalent on real inputs (the only case where ``scale_safe`` differs
  from ``scale`` is ``global_amax == 0``, in which case ``w_abs`` is also
  zero so the loss is zero either way), but more consistent.
- Extract ``fp8_scale_candidates`` to a triton-free module
  ``_fp8_scale_candidates.py`` so the calibrator's reference sweep and the
  Triton kernel wrapper share one definition. Removes the duplicate copy
  in ``NVFP4MSECalibrator._generate_candidates``.
- Parity test: extend ``test_parity_random_weights`` to cover bf16 and
  fp16 in addition to fp32 by parametrizing on dtype, so the canonical
  parity grid (3 seeds × 3 num_blocks) is now exercised on every supported
  dtype. Folded the smaller ``test_parity_dtypes`` into this since it was
  a strict subset.
